refactor: delete local provider
