+++ lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:aimimi/models/user.dart';
-import 'package:aimimi/providers/goals_provider.dart';
 import 'package:aimimi/services/goal_service.dart';
 import 'package:aimimi/constants/styles.dart';
 import 'package:aimimi/views/login_view.dart';
@@ -36,9 +35,6 @@ class MyApp extends StatelessWidget {
       // home: MainView(),
       home: MultiProvider(
         providers: [
-          ChangeNotifierProvider<GoalsProvider>(
-            create: (_) => GoalsProvider(),
-          ),
           StreamProvider<List<Goal>>.value(
             initialData: [],
             value: GoalService().goals,
